home adjust

+++ ft_memset.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_memset.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: antgarci <antgarci@student.42malaga.c      +#+  +:+       +#+        */
+/*   By: codespace <codespace@student.42.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/04/20 15:18:15 by antgarci          #+#    #+#             */
-/*   Updated: 2026/04/21 12:06:26 by antgarci         ###   ########.fr       */
+/*   Updated: 2026/04/28 20:13:07 by codespace        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memset(void *str, int c, size_t n)
 {
-	int	counter;
+	size_t	counter;
 	unsigned char	*pin;
 
 	counter = 0;
